requirements.txt
bcrypt==4.2.0
cffi==1.17.1
cryptography==43.0.1
decorator==5.1.1
Deprecated==1.2.14
fabric==3.2.2
invoke==2.2.0
paramiko==3.4.1
pycparser==2.22
PyNaCl==1.5.0
python-dotenv==1.0.1
PyYAML==6.0.2
sshclient==0.1
wrapt==1.16.0
